serverless/app/config_loader.py
import os
import yaml

BASE_DIR = os.path.dirname(os.path.abspath(__file__))
config_path = os.path.join(BASE_DIR, '../config/service-config.yml')


def load_config(config_file_path):
    with open(config_file_path, 'r') as stream:
        try:
            return yaml.safe_load(stream)
        except yaml.YAMLError as exc:
            print(exc)
            return None


config = load_config(config_path)
