serverless/requirements.txt
boto3==1.34.137
fast-ulid==0.2.0
Flask==3.0.3
Flask-Cognito==1.21
Cerberus==1.3.5
firebase-admin==6.5.0
PyYAML==6.0.1
Werkzeug==3.0.3
